pyproject.toml
[project]
name = "tdd-project"
version = "0.1.0"
description = ""
authors = [
    {name = "danalmeidap",email = "dbatistadealmeida23@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "pytest-asyncio (>=1.1.0,<2.0.0)",
    "bson (>=0.5.10,<0.6.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
taskipy = "^1.14.1"
ruff = "^0.12.11"
pytest-asyncio = "^1.1.0"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']


[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'


[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = [
    "--strict-config",
    "--strict-markers",
    "--ignore=docs_src",
    '-p no:warnings'
]
pythonpath = "."

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev tdd_project/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=tdd_project -vv'
post_test = 'coverage html'